Viewing Microfinance in Africa through Text Mining
"Microfinance is regarded as an instrument for poverty alleviation in developed and especially developing geographies. For the purpose of this project we have picked two extensive reports on the state of Microfinance in Africa. The multiple-page reports recount the evolution of the micro-finance industry from its birth to its recent boom thanks to micro-lending platforms leveraging new technologies to provide loans to those in need. Our goal in this Text Mining project will be to determine, per the content of the reports, who might benefit the most, what are the main benefits of micro-loans and what are they used for, what is the global perspective on micro-loans – is it positive, negative or somewhere in the middle? Is it really an effective way to address financial distress or poverty? We will also provide an actionable business recommendation based on the findings and additional research gathered." - from the report